t2

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -1,12 +1,12 @@
 #include "main.h"
 #include <stdio.h>
 
-/*
- * _strlen - returns length of a string
+/**
+ * _strlen - Returns length of a string
  * @str: string to get length
  * Return: string length
  */
-size_t _strlen(const char *str)
+int _strlen(const char *str)
 {
 	size_t l = 0;
